src: Update UDP header length field after mangling
One would expect nfq_udp_mangle_ipv4() to take care of the length field
in the UDP header but it did not.  With this patch, it does. This patch
is very unlikely to adversely affect any existing userspace software
(that did its own length adjustment), because UDP checksumming was
broken.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>