src: Make sure pktb_alloc() works for IPv6 over AF_BRIDGE
At least on the local interface, the MAC header of an IPv6 packet specifies
IPv6 protocol (rather than IP). This surprised me, since the first octet of
the IP datagram is the IP version, but I guess it's an efficiency thing.

Without this patch, pktb_alloc() returns NULL when an IPv6 packet is
encountered.

Updated:

 src/extra/pktbuff.c: - Treat ETH_P_IPV6 the same as ETH_P_IP.

Signed-off-by: Duncan Roe <duncan_roe@optusnet.com.au>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>