build: doc: Allow to specify whether to produce man pages, html, neither or both
New default action is: run doxygen (if installed) to produce man pages only.
This adds 124 KB to the build tree (and to the install tree, after
`make install`).
For finer control of built documentation, the old --with-doxygen configure
option is removed. Instead there are 2 new options:
  --enable-html-doc      # +1160 KB
  --disable-man-pages    #  -124 KB
If doxygen is not installed, configure outputs a warning that man pages will not
be built. configure --disable-man-pages avoids this warning.
After --enable-html-doc
 - `make install` installs built pages in htmldir instead of just leaving them
   in the build tree.
 - If the 'dot' program is not installed, configure outputs a warning that
   interactive diagrams will be missing and to install graphviz to get them.
   (There is an interactive diagram at the head of some modules, e.g.
    User-space network packet buffer).

[ a few configure.ac edits to keep it in sync with libnetfilter_log --pablo]

Signed-off-by: Duncan Roe <duncan_roe@optusnet.com.au>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>