build: doc: Warn user if html docs will be missing diagrams
libnetfilter_queue is unique among the netfilter libraries in having a
module hierarchy.

If 'dot' is available, Doxygen will make an interactive diagram for a
module with a child or a parent, allowing users to conveniently move up
and down the hierarchy.

Update configure to output a warning if 'dot' is not installed and html
was requested.

Signed-off-by: Duncan Roe <duncan_roe@optusnet.com.au>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>